Refactor frontend structure with routing
