+++ frontend/package.json
@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "axios": "^1.8.4",
-    "cra-template": "1.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-router-dom": "^7.5.1",
-    "react-scripts": "5.0.1"
+    "react-router-dom": "^6.22.2",
+    "react-scripts": "5.0.1",
+    "pdfmake": "^0.2.7"
   },
   "scripts": {
     "start": "craco start",
