+++ frontend/src/components/EventModal.jsx
@@ -1,0 +1,2 @@
+import { EventModal } from '../LegacyApp';
+export default EventModal;
