+++ frontend/src/components/InvoiceModal.jsx
@@ -1,0 +1,2 @@
+import { InvoiceModal } from '../LegacyApp';
+export default InvoiceModal;
