+++ frontend/src/components/QuoteModal.jsx
@@ -1,0 +1,2 @@
+import { QuoteModal } from '../LegacyApp';
+export default QuoteModal;
