+++ frontend/src/components/RevenueCards.jsx
@@ -1,0 +1,2 @@
+import { RevenueCards } from '../LegacyApp';
+export default RevenueCards;
