+++ frontend/src/components/TaskModal.jsx
@@ -1,0 +1,2 @@
+import { TaskModal } from '../LegacyApp';
+export default TaskModal;
