+++ frontend/src/components/WeekNavigationHeader.jsx
@@ -1,0 +1,2 @@
+import { WeekNavigationHeader } from '../LegacyApp';
+export default WeekNavigationHeader;
