+++ frontend/src/pages/Dashboard.jsx
@@ -1,0 +1,2 @@
+import { Dashboard } from '../LegacyApp';
+export default Dashboard;
