+++ frontend/src/pages/Invoices.jsx
@@ -1,0 +1,2 @@
+import { Invoices } from '../LegacyApp';
+export default Invoices;
