+++ frontend/src/pages/NotFound.jsx
@@ -1,0 +1,8 @@
+export default function NotFound() {
+  return (
+    <div className="p-6 text-center">
+      <h1 className="text-2xl font-bold mb-2">404 - Page not found</h1>
+      <p>Sorry, this page does not exist.</p>
+    </div>
+  );
+}
