+++ frontend/src/pages/Planning.jsx
@@ -1,0 +1,2 @@
+import { Planning } from '../LegacyApp';
+export default Planning;
