+++ frontend/src/pages/Quotes.jsx
@@ -1,0 +1,2 @@
+import { Quotes } from '../LegacyApp';
+export default Quotes;
