+++ frontend/src/utils/pdf.js
@@ -1,0 +1,2 @@
+export function generateQuotePDF() {}
+export function generateInvoicePDF() {}
